+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-### Unreleased
-- Fns-pipeline - first function for POC: Receiver-Debatcher
+## [0.0.3] - 2022-09-22
+- HL7 Structural Validator: added PHIN 2.0 and 3.0 profiles, updated to validate based on profile requested
+- Added Phin Spec 2.0 and 3.0 IGAMT Profiles.
+- HL7 validation function to structure validate messages
+- Started on HL7 Message Processor function
+- Updated Phin VADS function to be based on schedule.
+- Started on MMG Loader function.
 
 ## [0.0.2] - 2022-09-12 
  - Developed Receiver-Debatcher function
