+++ fns-hl7-pipeline/fn-message-processor/README.md
@@ -1,0 +1,13 @@
+
+## Message-Processor Function for the HL7 Pipeline
+
+
+### Function project was started with:
+
+[microsoft kotlin-maven](https://docs.microsoft.com/en-us/azure/azure-functions/functions-create-first-kotlin-maven?tabs=bash )
+
+```
+mvn archetype:generate \
+    -DarchetypeGroupId=com.microsoft.azure \
+    -DarchetypeArtifactId=azure-functions-kotlin-archetype
+```
