+++ fns-hl7-pipeline/fn-message-processor/function/host.json
@@ -1,0 +1,7 @@
+{
+  "version": "2.0",
+  "extensionBundle": {
+      "id": "Microsoft.Azure.Functions.ExtensionBundle",
+      "version": "[3.3.0, 4.0.0)"
+  }
+}
