+++ fns-hl7-pipeline/fn-message-processor/function/src/main/kotlin/rdmp/HL7Message.kt
@@ -1,0 +1,12 @@
+package com.example
+
+import com.google.gson.annotations.SerializedName
+
+
+data class HL7Message (
+
+  @SerializedName("content"  ) var content  : String,
+  @SerializedName("metadata" ) var metadata : HL7MessageMetadata,
+  // TODO:
+
+)
