+++ fns-hl7-pipeline/fn-receiver-debatcher/function/src/main/kotlin/rdmp/EvHubStorageDiagnostics.kt
@@ -1,0 +1,10 @@
+package com.example
+
+import com.google.gson.annotations.SerializedName
+
+
+data class EvHubStorageDiagnostics (
+
+  @SerializedName("batchId" ) var batchId : String
+
+)
