--- fns-hl7-pipeline/fn-receiver-debatcher/function/src/main/kotlin/receiverDebatcher/EvHubStorageDiagnostics.kt
@@ -1,10 +1,0 @@
-package com.example
-
-import com.google.gson.annotations.SerializedName
-
-
-data class EvHubStorageDiagnostics (
-
-  @SerializedName("batchId" ) var batchId : String
-
-)
