--- fns-hl7-pipeline/fn-receiver-debatcher/function/src/main/kotlin/receiverDebatcher/HL7Message.kt
@@ -1,11 +1,0 @@
-package com.example
-
-import com.google.gson.annotations.SerializedName
-
-
-data class HL7Message (
-
-  @SerializedName("content"  ) var content  : String,
-  @SerializedName("metadata" ) var metadata : HL7MessageMetadata,
-
-)
