+++ fns-hl7-pipeline/fn-structural-validator/README.md
@@ -1,0 +1,15 @@
+# Function HL7 Messages Structural Validation
+
+## Example install local dependency 
+
+- .jar available in /hl7-structural-validator
+
+```
+mvn install:install-file \
+    -Dfile=cdc.xlr.hl7.structurevalidator-assembly-0.4.4.jar \
+    -DgroupId=cdc.xlr.structurevalidator \
+    -DartifactId=cdc-xlr-hl7-structurevalidator \
+    -Dversion=0.4.4 \
+    -Dpackaging=jar \
+    -DgeneratePom=true
+```
