+++ hl7-structural-validator/src/main/scala/StructureValidatorConc.scala
@@ -9,7 +9,7 @@ import hl7.v2.validation.HL7Validator
 import gov.nist.validation.report.{Report}
 
 
-class StructureValidatorConc(profileLoader: ProfileLoader) extends StructureValidator {
+class StructureValidatorAsync(profileLoader: ProfileLoader) extends StructureValidator {
 
     val profile = profileLoader.profile()
     val valueSets = profileLoader.valueSets()
@@ -17,18 +17,17 @@ class StructureValidatorConc(profileLoader: ProfileLoader) extends StructureVali
 
     val validator = new HL7Validator(profile = profile, valueSetLibrary = valueSets, conformanceContext = confContext)
 
-        // returns Future[Map[String, Any]]  the JSON validation report or the error
+    // returns Future[Map[String, Any]]  the JSON validation report or the error
     def report(hl7Message: String): Try[Report] = {
 
       val reportFuture: Future[Report] = validator.validate(hl7Message, validator.profile.messages.keySet.head) 
       Try( Await.result(reportFuture, Duration(1, TimeUnit.SECONDS)) ) 
     } // .report
 
-} // .StructureValidatorConc
+} // .StructureValidatorAsync
 
 
-object StructureValidatorConc {
+object StructureValidatorAsync {
     
-  def apply() = new StructureValidatorConc(ProfileLoaderLocal())
-  def apply(profileLoader: ProfileLoader) = new StructureValidatorConc(profileLoader)
-} // .StructureValidatorConc
+  def apply(profileLoader: ProfileLoader) = new StructureValidatorAsync(profileLoader)
+} // .StructureValidatorAsync
