+++ hl7-structural-validator/src/main/scala/StructureValidatorSync.scala
@@ -33,7 +33,6 @@ class StructureValidatorSync(profileLoader: ProfileLoader) extends StructureVali
 
 object StructureValidatorSync {
 
-  def apply() = new StructureValidatorSync(ProfileLoaderLocal())
   def apply(profileLoader: ProfileLoader) = new StructureValidatorSync(profileLoader)
 } // .StructureValidatorSync
