+++ hl7-structural-validator/src/main/scala/package.scala
@@ -5,11 +5,11 @@ package object structurevalidator {
     val HL7_TEST_MESSAGES_LOCATION = "src/main/resources/HL7TestMessages/"
 
     // profiles available local
-    val PROFILES_PHIN_SPEC_3_1 = "PHIN Spec 3.1"
+    val PROFILES_PHIN_SPEC_2_0 = "phin_spec_2_0"
+    val PROFILES_PHIN_SPEC_3_0 = "phin_spec_3_0"
+    val PROFILES_PHIN_SPEC_3_1 = "phin_spec_3_1"
+    val PROFILES_PHIN_SPEC_3_2 = "phin_spec_3_2"
 
-
-    // Default profile
-    val PROFILES_LOCAL_PHIN_SPEC_3_1_FOLDER = "/phin_spec_3_1/"
     // Default profile file names
     val PROFILES_CONSTRAINTS_DEFAULT_FILE_NAME = "Constraints.xml"
     val PROFILES_PROFILE_DEFAULT_FILE_NAME = "Profile.xml"
